pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dataproc"
version = "0.1.0"
description = "An open-source data preprocessing library."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "Pillow>=10.0.0",
]
authors = [{ name = "dataproc team" }]

[project.urls]
Homepage = "https://github.com/327810854/dataproc"

[tool.setuptools.packages.find]
where = ["dataproc"]
include = ["dataproc*"]
